# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Show all services",
                "operationId": "4c175c5d6594b4a5112bbe15a46e2bb4",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "services": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/show/{service_id}": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Get service by id",
                "operationId": "c9bbb81ee35bf04820201e070ba196d3",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/show/section/{section_id}": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Get all services in section",
                "operationId": "d6c7a2d79c5cf3b5b5397fff681c0eaa",
                "parameters": [
                    {
                        "name": "section_id",
                        "in": "path",
                        "description": "ID of the section",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "allemployee": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No services found in this section",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No employee found in this service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/create": {
            "post": {
                "tags": [
                    "services"
                ],
                "summary": "create a new service",
                "operationId": "e5517ad7c6d1995a79a3055f358a45f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of service",
                                        "type": "string"
                                    },
                                    "desc": {
                                        "description": "service description",
                                        "type": "string"
                                    },
                                    "section_id": {
                                        "description": "section id",
                                        "type": "integer"
                                    },
                                    "image": {
                                        "description": "service image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "comment": {
                                        "description": "to access image use https://mahllola.online/public/image  example : https://mahllola.online/public/storage/services_folder/ttyVNuauz67kqXX40jyewMwh3DWpdFjjyJ0pIiPd.jpg",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "add new service successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/edit/{service_id}": {
            "post": {
                "tags": [
                    "services"
                ],
                "summary": "edit to services",
                "operationId": "b573ddab475963d46ba3b1184f896737",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "service name",
                                        "type": "string"
                                    },
                                    "desc": {
                                        "description": "service description",
                                        "type": "string"
                                    },
                                    "section_id": {
                                        "description": "section id",
                                        "type": "integer"
                                    },
                                    "image": {
                                        "description": "Image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "update service successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/delete/{service_id}": {
            "delete": {
                "tags": [
                    "services"
                ],
                "summary": "Delete an service",
                "description": "Delete service by ID",
                "operationId": "521513b2864d72e5a241e452726d3092",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "ID of the service to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "service deleted successfully"
                    },
                    "404": {
                        "description": "service not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}